Added when then end keywords for decision control.
Added when then end keywords for decision control.
Example Ascal Code for a recursive summation series function:
let f = when x>1 then f(x-1)+x when x<1 + x=1 then 1 end;
If the parser sees the keyword when it will evaluate the expression between the when, and the proceeding then. If the expression evaluates to anything other than 0 it will be interpreted as true, and the expression following the then will be executed, otherwise the next expression proceeding the next when clause will be executed, all cases must be covered, missing cases will cause the program to fail, working on else support.